requirements.txt
boto3==1.34.138
Django==5.0.7
django_crispy_forms==2.1
crispy-bootstrap5==2024.2
holidays==0.82
openai==2.20.0
openpyxl==3.1.2
pandas==3.0.0
Pillow==12.1.1
python-dotenv==1.2.1
python_dateutil==2.9.0.post0
pytz==2024.1
qrcode==7.4.2
reportlab==4.0.9
Requests==2.32.5
xhtml2pdf==0.2.15
djangorestframework==3.16.1
djangorestframework-api-key==3.0.0
django-storages==1.14.6
django-import-export==4.4.0
django-debug-toolbar==4.4.6
django-allauth[socialaccount]==0.63.1
psycopg[binary]==3.1.18
PyPDF2==3.0.1